+++ pkg/command/pretty_formatter.go
@@ -1,4 +1,4 @@
-package client
+package command
 
 import (
 	"bytes"
